crow-cli/pyproject.toml
[project]
name = "crow-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "agent-client-protocol>=0.8.1",
    "coolname>=2.2.0",
    "directory-tree>=1.0.0",
    "fastmcp>=2.14.5",
    "httpx>=0.28.1",
    "ipython>=9.10.0",
    "jinja2>=3.1.6",
    "json-schema-to-pydantic>=0.4.9",
    "openai>=2.21.0",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pysqlite3>=0.6.0",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "sqlalchemy>=2.0.46",
    "typer>=0.24.1",
]


[project.scripts]
crow-cli = "crow_cli.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crow_cli"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --strict-markers"
